+++ CMakeLists.txt
@@ -14,7 +14,7 @@ add_subdirectory(src/digest)
 add_subdirectory(src/generate)
 add_subdirectory(src/sign)
 add_subdirectory(src/encrypt)
-add_subdirectory(src/wrapping)
+add_subdirectory(src/wrap)
 add_subdirectory(src/hsm_info)
 add_subdirectory(src/attributes)
 add_subdirectory(src/destroy)
