+++ src/wrap/CMakeLists.txt
@@ -1,0 +1,28 @@
+cmake_minimum_required(VERSION 2.8)
+project(wrap)
+
+find_library(cloudhsmpkcs11 STATIC)
+
+add_executable(aes_gcm_wrap aes_gcm_wrap.c aes_wrap_common.c)
+add_executable(aes_no_padding_wrap aes_no_padding_wrap.c aes_wrap_common.c)
+add_executable(aes_zero_padding_wrap aes_zero_padding_wrap.c aes_wrap_common.c)
+add_executable(aes_pkcs5_padding_wrap aes_pkcs5_padding_wrap.c aes_wrap_common.c)
+add_executable(rsa_wrap rsa_wrap.c)
+add_executable(wrap_with_template wrap_with_template.c)
+add_executable(unwrap_with_template unwrap_with_template.c)
+
+target_link_libraries(aes_gcm_wrap cloudhsmpkcs11)
+target_link_libraries(aes_no_padding_wrap cloudhsmpkcs11)
+target_link_libraries(aes_zero_padding_wrap cloudhsmpkcs11)
+target_link_libraries(aes_pkcs5_padding_wrap cloudhsmpkcs11)
+target_link_libraries(rsa_wrap cloudhsmpkcs11)
+target_link_libraries(wrap_with_template cloudhsmpkcs11)
+target_link_libraries(unwrap_with_template cloudhsmpkcs11)
+
+add_test(aes_gcm_wrap aes_gcm_wrap --pin ${HSM_USER}:${HSM_PASSWORD})
+add_test(aes_no_padding_wrap aes_no_padding_wrap --pin ${HSM_USER}:${HSM_PASSWORD}) 
+add_test(aes_zero_padding_wrap aes_zero_padding_wrap --pin ${HSM_USER}:${HSM_PASSWORD})
+add_test(aes_pkcs5_padding_wrap aes_pkcs5_padding_wrap --pin ${HSM_USER}:${HSM_PASSWORD})
+add_test(rsa_wrap rsa_wrap --pin ${HSM_USER}:${HSM_PASSWORD}) 
+add_test(wrap_with_template wrap_with_template --pin ${HSM_USER}:${HSM_PASSWORD})
+add_test(unwrap_with_template unwrap_with_template --pin ${HSM_USER}:${HSM_PASSWORD} --wp_key ${TRUSTED_WRAPPING_KEY_HANDLE})
