--- src/wrapping/wrap.c
@@ -1,68 +1,0 @@
-/*
- * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-#include <unistd.h>
-#include <fcntl.h>
-#include "wrap.h"
-
-int main(int argc, char **argv) {
-    CK_RV rv;
-    CK_SESSION_HANDLE session;
-    int rc = 1;
-
-    struct pkcs_arguments args = {};
-    if (get_pkcs_args(argc, argv, &args) < 0) {
-        return rc;
-    }
-
-    rv = pkcs11_initialize(args.library);
-    if (CKR_OK != rv) {
-        return rc;
-    }
-
-    rv = pkcs11_open_session(args.pin, &session);
-    if (CKR_OK != rv) {
-        return rc;
-    }
-
-    printf("Running RSA wrap with OAEP padding...\n");
-    rc = rsa_oaep_wrap(session);
-    if (CKR_OK != rc) {
-        return rc;
-    }
-
-    printf("Running RSA AES wrap...\n");
-    rc = rsa_aes_wrap(session);
-    if (CKR_OK != rc) {
-        return rc;
-    }
-
-    printf("Running AES wrap...\n");
-    rc = aes_wrap(session);
-    if (CKR_OK != rc) {
-        return rc;
-    }
-
-    printf("Running AES-GCM wrap...\n");
-    rc = aes_gcm_wrap(session);
-    if (CKR_OK != rc) {
-        return rc;
-    }
-
-    pkcs11_finalize_session(session);
-    return rc;
-}
