--- src/wrapping/wrap.h
@@ -1,44 +1,0 @@
-/*
- * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-#ifndef AWS_CLOUDHSM_PKCS11_WRAP_H
-#define AWS_CLOUDHSM_PKCS11_WRAP_H
-
-#include <stdio.h>
-#include <stdint.h>
-#include <stdlib.h>
-#include <string.h>
-#include "common.h"
-
-int rsa_aes_wrap(CK_SESSION_HANDLE session);
-int rsa_oaep_wrap(CK_SESSION_HANDLE session);
-int aes_wrap(CK_SESSION_HANDLE session);
-int aes_gcm_wrap(CK_SESSION_HANDLE session);
-CK_RV rsa_aes_unwrap_key(
-        CK_SESSION_HANDLE session,
-        CK_OBJECT_HANDLE wrapping_key,
-        CK_KEY_TYPE wrapped_key_type,
-        CK_BYTE_PTR wrapped_bytes,
-        CK_ULONG wrapped_bytes_len,
-        CK_OBJECT_HANDLE_PTR unwrapped_key_handle);
-CK_RV rsa_aes_wrap_key(
-        CK_SESSION_HANDLE session,
-        CK_OBJECT_HANDLE wrapping_key,
-        CK_OBJECT_HANDLE key_to_wrap,
-        CK_BYTE_PTR wrapped_bytes,
-        CK_ULONG_PTR wrapped_bytes_len);
-
-#endif //AWS_CLOUDHSM_PKCS11_EXAMPLES_WRAP_H
